+++ src/shared/state/Position.h
@@ -17,8 +17,8 @@ namespace state {
     Position (int x = 0, int y = 0);
     void setX (int x);
     void setY (int y);
-    int const getX ();
-    int const getY ();
+    size_t const getX ();
+    size_t const getY ();
     // Setters and Getters
   };
